Remove unnecessary double conversion in PreloadingPrediction
The constructor takes the confidence as a double, but the caller
passes an int and the member is also an int.

Furthermore, an int64_t is being used despite knowing that the
possible values will be small. We switch to a plain int per the
style guide.

While we're here, we also remove some deprecated StringPiece usage.

Bug: None
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5371919
Auto-Submit: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1275617}

+++ chrome/browser/predictors/autocomplete_action_predictor.cc
@@ -339,7 +339,7 @@ AutocompleteActionPredictor::RecommendAction(
     // for logs.
     preloading_data->AddPreloadingPrediction(
         chrome_preloading_predictor::kOmniboxDirectURLInput,
-        static_cast<int64_t>(confidence * 100), std::move(same_url_matcher),
+        static_cast<int>(confidence * 100), std::move(same_url_matcher),
         web_contents->GetPrimaryMainFrame()->GetPageUkmSourceId());
   }
