+++ content/browser/preloading/preloading_data_impl.h
@@ -75,7 +75,7 @@ class CONTENT_EXPORT PreloadingDataImpl
       ukm::SourceId triggering_primary_page_source_id) override;
   void AddPreloadingPrediction(
       PreloadingPredictor predictor,
-      int64_t confidence,
+      int confidence,
       PreloadingURLMatchCallback url_match_predicate,
       ukm::SourceId triggering_primary_page_source_id) override;
   void SetIsNavigationInDomainCallback(
